# backend/complete_api_swagger.yaml
openapi: 3.0.3
info:
  title: School Management System - Complete API
  description: |
    Complete API documentation for School Management System with all endpoints.
    
    ## Features
    - Authentication & Authorization (JWT)
    - User Management (Students, Teachers, Parents, Admins)
    - Academic Management (Classes, Subjects, Terms)
    - Schedule & Attendance
    - Communication (Chat, Announcements, Meetings)
    - Teacher Performance Tracking
    - Task Management
    - Data Export (PDF, Excel, CSV)
    - Library Management
    - Leave Requests
    - Authentication Enhancements (OTP, 2FA)
    - Resource Requests
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolsystem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.schoolsystem.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        status:
          type: integer
          example: 200
        data:
          type: object
          description: Response data (varies by endpoint)
...
    TaskCreate:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: "Prepare lesson plan"
        description:
          type: string
          example: "Create detailed lesson plan for Math Grade 10"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "high"
        due_date:
          type: string
          format: date
          example: "2024-03-20"
        assigned_to:
          type: string
          format: uuid
    
    # Library Schemas
...
security:
  - bearerAuth: []
paths:
...
  /api/tasks/tasks/:
    get:
      tags:
        - Task Management
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, done, not_done]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: assigned_to
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "OK"
                status: 200
                data:
                  count: 25
                  results:
                    - id: "123e4567-e89b-12d3-a456-426614174000"
                      title: "Prepare lesson plan"
                      status: "in_progress"
                      priority: "high"
                      due_date: "2024-03-20"
    
    post:
      tags:
        - Task Management
      summary: Create new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              title: "Prepare lesson plan"
              description: "Create detailed lesson plan for Math Grade 10"
              priority: "high"
              due_date: "2024-03-20"
              assigned_to: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/tasks/tasks/{id}/:
    get:
      tags:
        - Task Management
      summary: Get task details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    put:
      tags:
        - Task Management
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [todo, in_progress, done, not_done]
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    delete:
      tags:
        - Task Management
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
...